The Loop is started and Finished!

+++ modules/brawl.js
@@ -2,14 +2,17 @@ const Enemies = require("./enemy.js");
 const Discord = require("discord.js");
 const Roll = require("../commands/roll.js");
 const Message = require("../events/message.js");
+const Loop = require("./loop.js");
+const Hbar = require("./hbar.js");
 
-exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
+exports.brawl = function(message, client, enemy1, playerFirst, color, dmg) {
 
   if (playerFirst){
     var damn = true;
       const embed = new Discord.RichEmbed()
         .setAuthor(message.author.username, message.author.avatarURL)
         .setColor(color)
+        .addField("HP", Hbar.hbar(enemy1,message,client))
         .addField("Enemy AV:",10+enemy1.agl)
         .addField("Your Turn!","Enter your damage if you hit, or press ➡ if you missed!")
         .addField("Form:",">[Damage Dealt]")
@@ -20,7 +23,6 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
       .then(sentEmbed => {
         sentEmbed.react("➡")
         .catch(() => console.log("fuk u"));
-
         const filter = (reaction, user) => {
           return['➡'].includes(reaction.emoji.name) && user.id === message.author.id;
         };
@@ -28,17 +30,23 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
         client.on('message', message => {
           if (message.content.indexOf(prefix) === 0 && damn) {
             const args = message.content.slice(prefix.length).trim().split(/ +/g);
-            message.channel.send('Test works I think? ' + args[0]);
-            sentEmbed.delete();
-            damn = false;
-            return
+            if (isNaN(args[0])){
+              message.channel.send("Try a number this time!");
+            }
+            else{
+              dmg=args[0];
+              Loop.loop(message, client, enemy1, true, true, color, dmg);
+              damn = false;
+              sentEmbed.delete();
           }
+            return
+        }
         });
         sentEmbed.awaitReactions(filter, {max: 1, time:60000, errors: ['time']})
           .then(collected => {
             const reaction = collected.first();
             if (reaction.emoji.name === '➡' && damn) {
-              message.channel.send('No, it doesn\'t!');
+              Loop.loop(message, client, enemy1, false, true, color, dmg);
               damn = false;
               sentEmbed.delete();
               return;
@@ -47,7 +55,7 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
               sentEmbed.delete();
           })
           .catch(collected => {
-            console.log ("someone fucked up something in Brawl-Player.");
+            console.log ("Missed Timed out, unlikely to be a real issue.");
           });
     });
   }
@@ -57,6 +65,7 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
       .setAuthor(message.author.username, message.author.avatarURL)
       .setColor(color)
       .addField("Enemy's Turn!",enemy1.type)
+      .addField("HP", Hbar.hbar(enemy1,message,client))
       .addField("Attack Roll:",Roll.roll(20) + enemy1.str)
       .addField("Does it hit you?","Hit Check for yes and X for no.")
       .setThumbnail(
@@ -67,7 +76,7 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
 
       sentEmbed.react("✅")
       .then(() => sentEmbed.react("❌"))
-      .catch(() => console.log("fuk u"));
+      .catch(() => console.log("fuk u 2"));
 
       const filter = (reaction, user) => {
         return['✅', '❌'].includes(reaction.emoji.name) && user.id === message.author.id;
@@ -78,19 +87,10 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
         const reaction = collected.first();
 
         if (reaction.emoji.name === '✅') {
-          const embed = new Discord.RichEmbed()
-             .setAuthor(message.author.username, message.author.avatarURL)
-             .setColor(color)
-             .addField("Damage Dealt to you:",Roll.roll(enemy1.bd)+enemy1.sd)
-             .addField("Your options are:","Abscond and Fight")
-             .setThumbnail(
-               (client.emojis.find
-                 (emoji => emoji.name === (enemy1.type).toLowerCase()).url));
-          message.channel.send({embed});
-
+        Loop.loop(message, client, enemy1, true, false, color, dmg);
         }
         else {
-          return;
+          Loop.loop(message, client, enemy1, false, false, color, dmg);
         }
         sentEmbed.delete();
       })
@@ -98,6 +98,5 @@ exports.brawl = function(message, client, enemy1, playerFirst, color, choice) {
         console.log ("someone fucked up something in Brawl-Enemy.");
       });
     });
-
   }
 };
